feat: readme + refactoring batch class package
